requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
chardet==3.0.4
discord.py==0.16.12
idna==2.7
idna-ssl==1.1.0
lxml==4.2.5
multidict==4.4.2
ujson==1.35
websockets==3.4
yarl==1.2.6
